add command create course

+++ src/main/java/by/epamtc/final_task/service/ResultUserService.java
@@ -1,15 +1,14 @@
 package by.epamtc.final_task.service;
 
-import by.epamtc.final_task.entity.Course;
 import by.epamtc.final_task.entity.ResultUser;
 import by.epamtc.final_task.service.exception.ServiceException;
 
-import java.util.List;
-
 public interface ResultUserService {
 
-    void addCourseResult(int studentId, int courseId, int mark, String comment) throws ServiceException;
-    ResultUser findResultUser(long userId,long courseId) throws ServiceException;
-    boolean updateUserCourseStatus(long userId,long courseId, ResultUser.UserCourseStatus status) throws ServiceException;
+    void addCourseResult(long studentId, long courseId, int mark, String comment) throws ServiceException;
+
+    ResultUser findResultUser(long userId, long courseId) throws ServiceException;
+
+    boolean updateUserCourseStatus(long userId, long courseId, ResultUser.UserCourseStatus status) throws ServiceException;
 
 }
